ci: fix the Ruby version parsing
Testing with Ruby 3.2 brings us to Ruby 3.2.8 which ships:

```sh
$ docker run -it ruby:3.2.8-alpine bundler --version
Bundler version 2.4.19
```

This is the version that introduces[1] the `ruby file:` syntax to read
the Ruby version from a file, but at the time it only handled a basic
"MAJOR.MINOR.PATCH" syntax, whereas our file has `ruby X.Y.Z` in it.

Support for this extended syntax came in v2.4.20[2] which we can only
get by dropping support for 3.2.

[1]: https://github.com/rubygems/rubygems/releases/tag/bundler-v2.4.19
[2]: https://github.com/rubygems/rubygems/releases/tag/bundler-v2.4.20